langflow_router/models/state.py
from typing_extensions import TypedDict, Annotated
from langgraph.graph.message import add_messages

class State(TypedDict):
    messages: Annotated[list, add_messages]
    options: list[str]
    route: str
    response: str
    greetings: bool