pyproject.toml
[project]
name = "langflow-router"
version = "1.0.0"
description = "Classification and routing layer for LangGraph agents"
authors = [{ name="Dário Ribeiro", email="root@darioribeiro.com" }]
dependencies = [
    "langgraph",
    "langchain",
    "langchain-aws",
    "pydantic",
    "python-dotenv"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
